+++ packages/zowe-explorer-ftp-extension/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to the "vscode-extension-for-zowe" extension will be documen
 
 ### New features and enhancements
 
-- Updated UI/UX method calls to use standalone `Gui` module for better usability and maintainability . [#1967](https://github.com/zowe/vscode-extension-for-zowe/issues/1967)
+- Added `putContent` to `FtpUssApi` for compatibility with optimizations to USS copy/paste actions.
 
 ### Bug fixes